requirements.txt
fastapi
uvicorn
jinja2
python-multipart
python-dotenv
sqlalchemy
email-validator
pydantic
